Improvements to test_discovery
Move everything into one file and make iteration test useful.
