+++ tests/test_cases/test_discovery/Makefile
@@ -28,8 +28,4 @@
 
 include ../../designs/sample_module/Makefile
 
-ifeq ($(TOPLEVEL_LANG), vhdl)
-MODULE = test_discovery,test_vhdl_indexed_name,test_vhdl_block
-else
 MODULE = test_discovery
-endif
